Allow dynamic selection of weather provider
